+++ inst/app/www/NEWS.md
@@ -6,6 +6,8 @@
 * Glycosylation sites are now ordered alphabetically throughout the dashboard
 and in the output data.
 * Make the download button for the user manual more obvious.
+* When spectra curation is skipped, the scatter plots with sum intensities
+and percentage of passing analytes will be shown in the HTML report.
 
 ## Bug fixes
 * Minor bug fixes.